rpdd: - Rewrite from MySQL C API to PostgreSQL C API (libpq)
The main reason for this insane task is that MySQL InnoDB has
row level locking but it's implementation sucks. There are so
many situations where they fall back to table locks that you
cannot use it in real life. This is unacceptable for all locks
that are kept for more then a second and we have quite a few
situations where we need exactly that behavior.